Move all migration logic to `sozo`
commit-id:a8445393
