+++ crates/dojo-world/src/lib.rs
@@ -1,3 +1,1 @@
-pub mod config;
 pub mod manifest;
-pub mod migration;
