+++ crates/sozo/Cargo.toml
@@ -36,6 +36,7 @@ indicatif = "0.17.4"
 
 [dev-dependencies]
 assert_fs = "1.0.10"
+dojo-test-utils = { path = "../dojo-test-utils" }
 snapbox = "0.4.6"
 tokio = { version = "1.28.0", features = [ "full" ] }
