+++ crates/sozo/src/ops/migration/mod.rs
@@ -1,3 +1,4 @@
+pub mod config;
 pub mod object;
 pub mod strategy;
 pub mod world;
