+++ crates/sozo/src/ops/mod.rs
@@ -1,0 +1,1 @@
+pub mod migration;
